backend/requirements.txt
fastapi==0.104.1
starlette==0.27.0
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
loguru==0.7.2
sqlalchemy==2.0.23
alembic==1.12.1
uvicorn==0.24.0
asyncpg==0.29.0
psycopg2==2.9.9
celery==5.3.5
redis==5.0.1
asgiref==3.7.2
aiohttp==3.8.6
certifi==2023.7.22
python-dateutil==2.8.2
gunicorn==21.2.0
flask==3.0.0
flask-admin==1.6.1
Pillow==10.1.0
